+++ src/pages/en/faq.tsx
@@ -1,11 +1,11 @@
-import FAQ from '@/components/faqPage/FAQPage';
 import Layout from '@/components/layout/Layout';
 import { NavBarText_ENG, FooterText_ENG, FAQText_ENG } from '@/texts/en';
+import FAQTemplate from '@/components/templates/FAQTemplate';
 
 const faq = () => {
   return (
     <Layout navBarTexts={NavBarText_ENG} footerTexts={FooterText_ENG}>
-      <FAQ texts={FAQText_ENG} />
+      <FAQTemplate texts={FAQText_ENG} />
     </Layout>
   );
 };
