+++ src/pages/es/faq.tsx
@@ -1,11 +1,11 @@
-import FAQ from '@/components/faqPage/FAQPage';
 import Layout from '@/components/layout/Layout';
 import { NavBarText_ESP, FooterText_ESP, FAQText_ESP } from '@/texts/es';
+import FAQTemplate from '@/components/templates/FAQTemplate';
 
 const faq = () => {
   return (
     <Layout navBarTexts={NavBarText_ESP} footerTexts={FooterText_ESP}>
-      <FAQ texts={FAQText_ESP} />
+      <FAQTemplate texts={FAQText_ESP} />
     </Layout>
   );
 };
