+++ src/pages/faq.tsx
@@ -1,11 +1,11 @@
-import FAQ from '@/components/faqPage/FAQPage';
 import Layout from '@/components/layout/Layout';
 import { NavBarText_KR, FooterText_KR, FAQText_KR } from '@/texts/ko';
+import FAQTemplate from '@/components/templates/FAQTemplate';
 
 const faq = () => {
   return (
     <Layout navBarTexts={NavBarText_KR} footerTexts={FooterText_KR}>
-      <FAQ texts={FAQText_KR} />
+      <FAQTemplate texts={FAQText_KR} />
     </Layout>
   );
 };
