+++ src/pages/in/faq.tsx
@@ -1,11 +1,11 @@
-import FAQ from '@/components/faqPage/FAQPage';
 import Layout from '@/components/layout/Layout';
 import { NavBarText_IND, FooterText_IND, FAQText_IND } from '@/texts/in';
+import FAQTemplate from '@/components/templates/FAQTemplate';
 
 const faq = () => {
   return (
     <Layout navBarTexts={NavBarText_IND} footerTexts={FooterText_IND}>
-      <FAQ texts={FAQText_IND} />
+      <FAQTemplate texts={FAQText_IND} />
     </Layout>
   );
 };
