+++ src/pages/ja/faq.tsx
@@ -1,11 +1,11 @@
-import FAQ from '@/components/faqPage/FAQPage';
 import Layout from '@/components/layout/Layout';
 import { NavBarText_JAP, FooterText_JAP, FAQText_JAP } from '@/texts/ja';
+import FAQTemplate from '@/components/templates/FAQTemplate';
 
 const faq = () => {
   return (
     <Layout navBarTexts={NavBarText_JAP} footerTexts={FooterText_JAP}>
-      <FAQ texts={FAQText_JAP} />
+      <FAQTemplate texts={FAQText_JAP} />
     </Layout>
   );
 };
