+++ src/pages/vi/faq.tsx
@@ -1,11 +1,11 @@
-import FAQ from '@/components/faqPage/FAQPage';
 import Layout from '@/components/layout/Layout';
 import { NavBarText_VIE, FooterText_VIE, FAQText_VIE } from '@/texts/vi';
+import FAQTemplate from '@/components/templates/FAQTemplate';
 
 const faq = () => {
   return (
     <Layout navBarTexts={NavBarText_VIE} footerTexts={FooterText_VIE}>
-      <FAQ texts={FAQText_VIE} />
+      <FAQTemplate texts={FAQText_VIE} />
     </Layout>
   );
 };
