+++ src/pages/zh-rTW/faq.tsx
@@ -1,11 +1,11 @@
-import FAQ from '@/components/faqPage/FAQPage';
 import Layout from '@/components/layout/Layout';
 import { NavBarText_zh_rTW, FooterText_zh_rTW, FAQText_zh_rTW } from '@/texts/zh-rTW';
+import FAQTemplate from '@/components/templates/FAQTemplate';
 
 const faq = () => {
   return (
     <Layout navBarTexts={NavBarText_zh_rTW} footerTexts={FooterText_zh_rTW}>
-      <FAQ texts={FAQText_zh_rTW} />
+      <FAQTemplate texts={FAQText_zh_rTW} />
     </Layout>
   );
 };
